fix: state encapsulation
