+++ src/flows/components/FlowPage.tsx
@@ -1,20 +1,42 @@
-import React from 'react'
+// Libraries
+import React, {FC, useContext, useEffect} from 'react'
+import {useParams} from 'react-router-dom'
 
 // Components
+import CurrentFlowProvider from 'src/flows/context/flow.current'
+import {FlowListContext} from 'src/flows/context/flow.list'
+import Flow from 'src/flows/components/Flow'
 import {Page} from '@influxdata/clockface'
 import FlowHeader from 'src/flows/components/header'
-import PipeList from 'src/flows/components/PipeList'
-import QueryProvider from 'src/flows/context/query'
+import {ResultsProvider} from 'src/flows/context/results'
 
-const FlowPage = () => (
-  <Page titleTag="Flows">
-    <FlowHeader />
-    <QueryProvider>
-      <Page.Contents fullWidth={true} scrollable={true} className="flow-page">
-        <PipeList />
-      </Page.Contents>
-    </QueryProvider>
-  </Page>
+const FlowFromRoute = () => {
+  const {id} = useParams()
+  const {change} = useContext(FlowListContext)
+
+  useEffect(() => {
+    change(id)
+  }, [id, change])
+
+  return null
+}
+// NOTE: uncommon, but using this to scope the project
+// within the page and not bleed it's dependencies outside
+// of the feature flag
+import 'src/flows/style.scss'
+
+const FlowContainer: FC = () => (
+  <CurrentFlowProvider>
+    <FlowFromRoute />
+    <ResultsProvider>
+      <Page titleTag="Flows">
+        <FlowHeader />
+        <Page.Contents fullWidth={true} scrollable={true} className="flow-page">
+          <Flow />
+        </Page.Contents>
+      </Page>
+    </ResultsProvider>
+  </CurrentFlowProvider>
 )
 
-export default FlowPage
+export default FlowContainer
